--- budget-app-main/src/App.css
@@ -1,91 +1,0 @@
-.App {
-  text-align: center;
-  height: 100vh;
-  display: flex;
-  flex-direction: column;
-  justify-content: flex-start;
-  overflow: hidden;
-}
-
-.BillPieSaving {
-  display: flex;
-  justify-content: space-between; 
-  align-items: flex-start;  
-  gap: 10px;  
-  padding: 20px;
-  max-width: 90%;
-  box-sizing: border-box;
-  flex-wrap: wrap;  
-}
-
-.BillPieSaving > * {
-  flex: 1 1 30%;  
-  min-width: 200px;  
-  box-sizing: border-box;
-}
-
-.Salary-Calculator {
-  display: flex;
-  flex-wrap: wrap; 
-  justify-content: space-between;
-  padding: 10px 20px; 
-  background-color: #85BB65;
-  font-family: Arial, Helvetica, sans-serif;
-  min-height: 35vh; 
-  box-sizing: border-box;
-}
-
-.Salary-Calculator .input-section,
-.Salary-Calculator .breakdown-section {
-  display: flex;
-  flex-direction: column;
-  gap: 8px; 
-  width: 48%; 
-}
-
-.Salary-Calculator input,
-.Salary-Calculator .breakdown-section p {
-  padding: 8px; 
-  margin: 0; 
-  font-size: 1rem; 
-  box-sizing: border-box; 
-}
-
-/* Savings Display */
-.SavingsDisplay {
-  width: 30%; 
-  padding: 20px;
-  background: linear-gradient(135deg, #42a5f5, #2979ff);
-  border-radius: 10px;
-  box-shadow: 0 6px 12px rgba(0, 0, 0, 0.15);
-  text-align: center;
-  color: white;
-  border: 2px solid rgba(255, 255, 255, 0.1);
-  min-width: 250px;
-}
-
-/* Media Queries */
-@media screen and (max-width: 768px) {
-  .Salary-Calculator {
-    flex-direction: column;
-    align-items: center;
-    height: auto; 
-  }
-
-  .BillPieSaving {
-    flex-direction: column;
-    align-items: center;
-    gap: 10px; 
-  }
-
-  .SavingsDisplay {
-    width: 90%;
-    margin-bottom: 20px; 
-  }
-
-  .Salary-Calculator .input-section,
-  .Salary-Calculator .breakdown-section {
-    width: 100%; 
-    gap: 5px; 
-  }
-}
