--- budget-app-main/src/banner.css
@@ -1,21 +1,0 @@
-.Banner-header {
-  background-color: #85BB65;
-  color: white;
-  padding: 10px;
-  font-size: 1.5rem;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  font-family: Arial, Helvetica, sans-serif;
-}
-
-.dollar-icon {
-  width: 30px;
-  height: auto;
-}
-
-.Banner-text {
-  color: #084f09;
-  font-weight: bold;
-  font-size: 3rem;
-}
