--- budget-app-main/src/bills.css
@@ -1,41 +1,0 @@
-.Bills {
-  padding-left: 20px;
-}
-
-.bill-list button {
-  background-color: #b00020;
-  color: white;
-  border: none;
-  border-radius: 4px;
-  padding: 4px 10px;
-  font-size: 0.9rem;
-  cursor: pointer;
-}
-
-.bill-list button:hover {
-  background-color: #d32f2f;
-}
-
-.input-row {
-  display: flex;
-  justify-content: center;
-  flex-wrap: wrap;
-  gap: 10px;
-  margin-top: 12px;
-  margin-bottom: 16px;
-}
-
-.input-row input,
-.input-row select {
-  background-color: #2a2a2a;
-  color: #ffffff;
-  border: 1px solid #444;
-  padding: 8px;
-  border-radius: 6px;
-  font-size: 1rem;
-  width: 180px;
-}
-
-.input-row input::placeholder {
-  color: #ffffff;
-}
