--- budget-app-main/src/components/SavingsDisplay.css
@@ -1,26 +1,0 @@
-.savings-label {
-  text-transform: uppercase;
-  letter-spacing: 3px;
-  font-size: 18px;
-  margin: 0 0 15px 0;
-  font-weight: 500;
-  text-shadow: 0 1px 2px rgba(0, 0, 0, 0.2);
-}
-
-.savings-amount {
-  font-size: 42px;
-  font-weight: 700;
-  margin: 0;
-  text-shadow: 0 2px 4px rgba(0, 0, 0, 0.2);
-  transition: all 0.3s ease;
-}
-
-.savings-amount:hover {
-  transform: scale(1.05);
-}
-
-@media (max-width: 768px) {
-  .savings-amount {
-    font-size: 36px;
-  }
-}
