--- budget-app-main/src/salary.css
@@ -1,41 +1,0 @@
-.Salary-input {
-  display: flex;
-  flex-direction: column;
-  width: 48%;
-  gap: 5px;
-}
-
-.Salary-result {
-  display: flex;
-  flex-direction: column;
-  width: 48%;
-  gap: 5px;
-}
-
-input {
-  padding: 8px;
-  font-size: 1rem;
-  border: 1px solid #ccc;
-  border-radius: 4px;
-  width: 100%;
-}
-
-label {
-  font-weight: bold;
-}
-
-button {
-  background-color: #000;
-  color: #fff;
-  border: none;
-  padding: 8px 16px;
-  border-radius: 6px;
-  font-size: 1rem;
-  cursor: pointer;
-  margin: 6px 6px 12px 0;
-  transition: background-color 0.2s ease;
-}
-
-button:hover {
-  background-color: #222;
-}
