+++ src/banner.css
@@ -1,22 +1,21 @@
-  .Banner-header {
-    background-color: #85BB65;
-    color: white;
-    padding: 10px;
-    text-align: center;
-    font-size: 1.5rem;
-    display: flex;
-    align-items: center;
-    justify-content: space-between;
-    font-family: Arial, Helvetica, sans-serif;
-  }
-  
-  .dollar-icon {
-    width: 30px;  /* Adjust size as needed */
-    height: auto; /* Maintain aspect ratio */
-  }
-  
-  .Banner-text {
-    color: #084f09; /* Sets text color */
-    font-weight: bold; /* Makes the text bold */
-    font-size: 3rem; /* Increases font size */
-}      
+.Banner-header {
+  background-color: #85BB65;
+  color: white;
+  padding: 10px;
+  font-size: 1.5rem;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  font-family: Arial, Helvetica, sans-serif;
+}
+
+.dollar-icon {
+  width: 30px;
+  height: auto;
+}
+
+.Banner-text {
+  color: #084f09;
+  font-weight: bold;
+  font-size: 3rem;
+}
