+++ src/components/SavingsDisplay.css
@@ -1,14 +1,3 @@
-.savings-container {
-  width: 100%;
-  padding: 20px;
-  background: linear-gradient(135deg, #42a5f5, #2979ff);
-  border-radius: 10px;
-  box-shadow: 0 6px 12px rgba(0, 0, 0, 0.15);
-  text-align: center;
-  color: white;
-  border: 2px solid rgba(255, 255, 255, 0.1);
-}
-
 .savings-label {
   text-transform: uppercase;
   letter-spacing: 3px;
