+++ src/salary.css
@@ -1,15 +1,41 @@
-.Salary-Calculator {
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-    padding-left: 20px;
-    padding-top: 10px;
-    height: 60vh;
-    background-color: #85BB65;
-    font-family: Arial, Helvetica, sans-serif;
-}
 .Salary-input {
-    color: #084f09; /* Sets text color */
-    font-weight: bold; /* Makes the text bold */
-    font-size: 1.25rem;
+  display: flex;
+  flex-direction: column;
+  width: 48%;
+  gap: 5px;
+}
+
+.Salary-result {
+  display: flex;
+  flex-direction: column;
+  width: 48%;
+  gap: 5px;
+}
+
+input {
+  padding: 8px;
+  font-size: 1rem;
+  border: 1px solid #ccc;
+  border-radius: 4px;
+  width: 100%;
+}
+
+label {
+  font-weight: bold;
+}
+
+button {
+  background-color: #000;
+  color: #fff;
+  border: none;
+  padding: 8px 16px;
+  border-radius: 6px;
+  font-size: 1rem;
+  cursor: pointer;
+  margin: 6px 6px 12px 0;
+  transition: background-color 0.2s ease;
+}
+
+button:hover {
+  background-color: #222;
 }
