+++ src/salary.js
@@ -1,11 +1,11 @@
-import { useState } from "react";
+import { useState, useEffect } from "react";
 import "./salary.css";
 
-const Salary = () => {
+const Salary = ({ setSalary }) => {
   const [yearlySalary, setYearlySalary] = useState("");
   const [hourlyRate, setHourlyRate] = useState("");
 
-  // Calculate salary based on input
+  // Calculate yearly salary based on input
   const calculatedYearlySalary = yearlySalary
     ? parseFloat(yearlySalary)
     : hourlyRate
@@ -16,8 +16,23 @@ const Salary = () => {
   const weeklySalary = (calculatedYearlySalary / 52).toFixed(2);
   const hourlySalary = (calculatedYearlySalary / (40 * 52)).toFixed(2); // Assuming 40 hours per week, 52 weeks per year
 
+  // Format function to add commas to salary amounts
+  const formatSalary = (amount) => {
+    return new Intl.NumberFormat("en-US", {
+      style: "decimal",
+      minimumFractionDigits: 2,
+      maximumFractionDigits: 2,
+    }).format(amount);
+  };
+
+  // Pass the calculated salary to the parent component (App.js)
+  useEffect(() => {
+    setSalary(calculatedYearlySalary);  // Update the salary in the parent component whenever it changes
+  }, [calculatedYearlySalary, setSalary]);
+
   return (
     <div className="Salary-Calculator">
+      {/* Salary Inputs Section */}
       <div className="Salary-input">
         <h2>Salary Breakdown</h2>
 
@@ -46,15 +61,15 @@ const Salary = () => {
           }}
           placeholder="Enter hourly rate"
         />
+      </div>
 
-        {/* Always Show Breakdown */}
-        <div className="Salary-result">
-          <h3>Salary Breakdown:</h3>
-          <p>Yearly: ${calculatedYearlySalary.toFixed(2)}</p>
-          <p>Monthly: ${monthlySalary}</p>
-          <p>Weekly: ${weeklySalary}</p>
-          <p>Hourly: ${hourlySalary}</p>
-        </div>
+      {/* Salary Breakdown Section */}
+      <div className="Salary-result">
+        <h3>Salary Breakdown:</h3>
+        <p>Yearly: ${formatSalary(calculatedYearlySalary)}</p>
+        <p>Monthly: ${formatSalary(monthlySalary)}</p>
+        <p>Weekly: ${formatSalary(weeklySalary)}</p>
+        <p>Hourly: ${formatSalary(hourlySalary)}</p>
       </div>
     </div>
   );
